Fixed some errors and improved logging

+++ tonumove.py
@@ -3,6 +3,21 @@
 import argparse
 import os
 import shutil
+import logging
+
+
+def setUpLogger(logfile):
+    logger = logging.getLogger('Tonumove')
+    logger.setLevel(logging.INFO)
+    fh = logging.FileHandler(logfile, 'w')
+    ch = logging.StreamHandler()
+    formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+    fh.setFormatter(formatter)
+    ch.setFormatter(formatter)
+    # add the handlers to the logger
+    logger.addHandler(fh)
+    logger.addHandler(ch)
+    return logger
 
 
 def main():
@@ -25,18 +40,22 @@ def main():
     parser.add_argument("-o", "--output", required=True, help='Folder where sd card is mounted', type=str)
     parser.add_argument("--superfolder", action='store_true', help='Input if folder which contains multiple folders or input files (each copied to a differnent destination folder')
     parser.add_argument("--overwrite", action='store_true', help='Clear sd-card before copying')
+    parser.add_argument("--logfile", default='tonumove.log', help='Log file', type=str)
 
     args = parser.parse_args()
+
+    logger = setUpLogger(args.logfile)
 
     if args.overwrite:
         answer = input("Are you sure you want to overwrite the sd card at %s? (y/n)" % args.output)
         if answer != 'y' and answer != 'Y':
-            print('Abort')
+            logger.info('Abort')
             return
         else:
-            print('Clear %s' % args.output)
+            logger.info('Clear %s' % args.output)
             for f in [folder for folder, _, _ in os.walk(args.output)][1:]:
-                shutil.rmtree(f)
+                if not f.endswith('advert') and not f.endswith('mp3'):
+                    shutil.rmtree(f)
 
     if not args.superfolder:
         copy2Tonuino(args.input, args.output)
@@ -44,11 +63,19 @@ def main():
         for folder, subfolders, files in os.walk(args.input):
             notRoot = folder is not args.input
             hasMp3 = len([f for f in os.listdir(folder) if f.endswith('.mp3')]) > 0
-            if notRoot and hasMp3:
-                print ('Processing %s' % folder)
+            if not notRoot and hasMp3:
+                for f in files:
+                    print(f)
+                    logger.info('Processing %s' % f)
+                    result = copy2Tonuino(os.path.join(folder, f), args.output)
+                    if result != 0:
+                        logger.error('Error while processing %s. Abort.' % f)
+                        return
+            elif notRoot and hasMp3:
+                logger.info('Processing %s' % folder)
                 result = copy2Tonuino(folder, args.output)
                 if result != 0:
-                    print('Error while processing %s. Abort.' % folder)
+                    logger.error('Error while processing %s. Abort.' % folder)
                     return
 
 
@@ -88,33 +115,36 @@ def fixFileOrder(files):
 
 
 def checkSanity(sdcard):
+    logger = logging.getLogger('Tonumove')
     for n, (folder, subfolders, files) in enumerate(os.walk(sdcard)):
         if n == 0:
             for s in subfolders:
-                if not len(s) == 2:
-                    print('Warning: Subfolder %s has length > 2' % s)
-                    return False
-                if not str.isdigit(s[0]) and str.isdigit(s[1]):
-                    print('Warning: Subfolder %s has non-digit characters' % s)
-                    return False
-                if not int(s) > 0 and not int(s) < 100:
-                    print('Warning: Subfolder %s number is not between 1 and 99' % s)
-                    return False
+                if not s.endswith('advert') and not s.endswith('mp3'):
+                    if not len(s) == 2:
+                        logger.warning('Subfolder %s has length > 2' % s)
+                        return False
+                    if not str.isdigit(s[0]) and str.isdigit(s[1]):
+                        logger.warning('Subfolder %s has non-digit characters' % s)
+                        return False
+                    if not int(s) > 0 and not int(s) < 100:
+                        logger.warning('Subfolder %s number is not between 1 and 99' % s)
+                        return False
         if n > 0:
-            for f in files:
-                if not len(f) == 7:
-                    print('Warning: File %s in folder %s does not have length 7' % (f, folder))
-                    return False
-                track = f[:3]
-                if not f.endswith('.mp3'):
-                    print('Warning: File %s in folder %s  does not end with .mp3' % (f, folder))
-                    return False
-                if not str.isdigit(track[0]) and str.isdigit(track[1]) and str.isdigit(track[2]):
-                    print('Warning: File %s in folder %s is not comprised of digits' % (f, folder))
-                    return False
-                if not int(track) > 0 and not int(track) < 256:
-                    print('Warning: File %s in folder %s is not digits between 1 and 255' % (f, folder))
-                    return False
+            if not folder.endswith('advert') and not folder.endswith('mp3'):
+                for f in files:
+                    if not len(f) == 7:
+                        logger.warning('File %s in folder %s does not have length 7' % (f, folder))
+                        return False
+                    track = f[:3]
+                    if not f.endswith('.mp3'):
+                        logger.warning('File %s in folder %s  does not end with .mp3' % (f, folder))
+                        return False
+                    if not str.isdigit(track[0]) and str.isdigit(track[1]) and str.isdigit(track[2]):
+                        logger.warning('File %s in folder %s is not comprised of digits' % (f, folder))
+                        return False
+                    if not int(track) > 0 and not int(track) < 256:
+                        logger.warning('File %s in folder %s is not digits between 1 and 255' % (f, folder))
+                        return False
     return True
 
 
@@ -125,6 +155,7 @@ def copy2Tonuino(input, output):
     input: file or folder with mp3 files
     output: mount point of sd card
     """
+    logger = logging.getLogger('Tonumove')
     # maximum number of folders on sd card
     maxNFolders = 100
     # maximum number of files in folders
@@ -137,11 +168,14 @@ def copy2Tonuino(input, output):
         ifiles = [f for _, f in sorted(zip(ifiles_fixed, ifiles))]
 
         if len(ifiles) == 0:
-            print("Error: No mp3 files found in folder %s" % input)
+            logger.error("No mp3 files found in folder %s" % input)
             return -1
         if len(ifiles) > maxNFiles:
-            print("Error: Too many (>%i) files in input folder folder %s" % (maxNFiles, input))
+            logger.error("Too many (>%i) files in input folder folder %s" % (maxNFiles, input))
             return -1
+
+        # get full path
+        ifiles = [os.path.join(input, f) for f in ifiles]
     else:
         ifiles = [input]
 
@@ -149,21 +183,20 @@ def copy2Tonuino(input, output):
     foldernames = ['%02i/' % n for n in range(1, maxNFolders)]
     foldersExists = [os.path.exists(os.path.join(output, name)) for name in foldernames]
     if all(foldersExists):
-        print("Error: All %i folders exists. Free up space by removing at least one folder" % maxNFolders)
+        logger.error("All %i folders exists. Free up space by removing at least one folder" % maxNFolders)
         return -1
     foldername = os.path.join(output, foldernames[foldersExists.index(False)])
     os.makedirs(foldername)
 
     # move files and rename them
-    print('Copying files')
+    logger.info('Copying files')
     for n, ifile in enumerate(ifiles):
-        src = os.path.join(input, ifile)
         dest = os.path.join(foldername, "%03i.mp3" % (n + 1))
-        print("%s->%s" % (src, dest))
-        shutil.copy(src, dest)
+        logger.info("%s->%s" % (ifile, dest))
+        shutil.copy(ifile, dest)
     # check for correct format
     if not checkSanity(output):
-        print("Warning content of sd-card does not meet format requirements. Please check")
+        logger.warning("Content of sd-card does not meet format requirements. Please check")
 
     return 0
